refactor(core): create dedicated type for numbers
