+++ harper-core/src/lib.rs
@@ -12,6 +12,7 @@ pub mod language_detection;
 mod lexing;
 pub mod linting;
 mod mask;
+mod number;
 pub mod parsers;
 pub mod patterns;
 mod punctuation;
@@ -34,13 +35,14 @@ pub use fat_token::FatToken;
 pub use ignored_lints::IgnoredLints;
 use linting::Lint;
 pub use mask::{Mask, Masker};
+pub use number::{Number, NumberSuffix};
 pub use punctuation::{Punctuation, Quote};
 pub use span::Span;
 pub use spell::{Dictionary, FstDictionary, FullDictionary, MergedDictionary};
 pub use sync::Lrc;
 pub use title_case::{make_title_case, make_title_case_str};
 pub use token::Token;
-pub use token_kind::{NumberSuffix, TokenKind};
+pub use token_kind::TokenKind;
 pub use token_string_ext::TokenStringExt;
 pub use vec_ext::VecExt;
 pub use word_metadata::{AdverbData, ConjunctionData, NounData, Tense, VerbData, WordMetadata};
