+++ harper-core/src/token_kind.rs
@@ -2,7 +2,7 @@ use is_macro::Is;
 use ordered_float::OrderedFloat;
 use serde::{Deserialize, Serialize};
 
-use crate::{ConjunctionData, NounData, Punctuation, Quote, WordMetadata};
+use crate::{ConjunctionData, NounData, Number, Punctuation, Quote, WordMetadata};
 
 #[derive(
     Debug, Is, Clone, Copy, Serialize, Deserialize, Default, PartialOrd, Hash, Eq, PartialEq,
@@ -11,7 +11,7 @@ use crate::{ConjunctionData, NounData, Punctuation, Quote, WordMetadata};
 pub enum TokenKind {
     Word(WordMetadata),
     Punctuation(Punctuation),
-    Number(OrderedFloat<f64>, Option<NumberSuffix>),
+    Number(Number),
     /// A sequence of " " spaces.
     Space(usize),
     /// A sequence of "\n" newlines
@@ -162,7 +162,7 @@ impl TokenKind {
         match self {
             TokenKind::Word(_) => TokenKind::Word(Default::default()),
             TokenKind::Punctuation(_) => TokenKind::Punctuation(Default::default()),
-            TokenKind::Number(..) => TokenKind::Number(Default::default(), Default::default()),
+            TokenKind::Number(..) => TokenKind::Number(Default::default()),
             TokenKind::Space(_) => TokenKind::Space(Default::default()),
             TokenKind::Newline(_) => TokenKind::Newline(Default::default()),
             _ => *self,
@@ -174,84 +174,6 @@ impl TokenKind {
     /// Construct a [`TokenKind::Word`] with no (default) metadata.
     pub fn blank_word() -> Self {
         Self::Word(WordMetadata::default())
-    }
-}
-
-#[derive(
-    Debug, Serialize, Deserialize, Default, PartialEq, PartialOrd, Clone, Copy, Is, Hash, Eq,
-)]
-pub enum NumberSuffix {
-    #[default]
-    Th,
-    St,
-    Nd,
-    Rd,
-}
-
-impl NumberSuffix {
-    pub fn correct_suffix_for(number: impl Into<f64>) -> Option<Self> {
-        let number = number.into();
-
-        if number < 0.0 || number - number.floor() > f64::EPSILON || number > u64::MAX as f64 {
-            return None;
-        }
-
-        let integer = number as u64;
-
-        if let 11..=13 = integer % 100 {
-            return Some(Self::Th);
-        };
-
-        match integer % 10 {
-            0 => Some(Self::Th),
-            1 => Some(Self::St),
-            2 => Some(Self::Nd),
-            3 => Some(Self::Rd),
-            4 => Some(Self::Th),
-            5 => Some(Self::Th),
-            6 => Some(Self::Th),
-            7 => Some(Self::Th),
-            8 => Some(Self::Th),
-            9 => Some(Self::Th),
-            _ => None,
-        }
-    }
-
-    pub fn to_chars(self) -> Vec<char> {
-        match self {
-            NumberSuffix::Th => vec!['t', 'h'],
-            NumberSuffix::St => vec!['s', 't'],
-            NumberSuffix::Nd => vec!['n', 'd'],
-            NumberSuffix::Rd => vec!['r', 'd'],
-        }
-    }
-
-    /// Check the first several characters in a buffer to see if it matches a
-    /// number suffix.
-    pub fn from_chars(chars: &[char]) -> Option<Self> {
-        if chars.len() < 2 {
-            return None;
-        }
-
-        match (chars[0], chars[1]) {
-            ('t', 'h') => Some(NumberSuffix::Th),
-            ('T', 'h') => Some(NumberSuffix::Th),
-            ('t', 'H') => Some(NumberSuffix::Th),
-            ('T', 'H') => Some(NumberSuffix::Th),
-            ('s', 't') => Some(NumberSuffix::St),
-            ('S', 't') => Some(NumberSuffix::St),
-            ('s', 'T') => Some(NumberSuffix::St),
-            ('S', 'T') => Some(NumberSuffix::St),
-            ('n', 'd') => Some(NumberSuffix::Nd),
-            ('N', 'd') => Some(NumberSuffix::Nd),
-            ('n', 'D') => Some(NumberSuffix::Nd),
-            ('N', 'D') => Some(NumberSuffix::Nd),
-            ('r', 'd') => Some(NumberSuffix::Rd),
-            ('R', 'd') => Some(NumberSuffix::Rd),
-            ('r', 'D') => Some(NumberSuffix::Rd),
-            ('R', 'D') => Some(NumberSuffix::Rd),
-            _ => None,
-        }
     }
 }
